backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
asyncpg==0.28.0
psycopg2-binary==2.9.7
databases[postgresql]==0.8.0
sqlalchemy==1.4.53
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Audio processing dependencies
aiofiles==23.2.1
librosa==0.10.1
soundfile==0.12.1
pydub==0.25.1
ffmpeg-python==0.2.0

# AI/ML dependencies
openai==1.3.0
google-generativeai==0.3.2
google-cloud-speech==2.21.0
transformers==4.35.2
torch==2.1.0